Fix double grant_type bug
The grant_type=authorization_code param is specified twice: in the URL
and in the request body. For some reason this was fine until yesterday,
but the OAuth endpoint policies must have become stricter, and now it
may be counting the param twice and therefore failing to complete the
auth process.

Change-Id: I0de0f57ae0ce80417795352ea9419297c296c414
Contributed-By: Anthony Weems <anthony.weems@praetorian.com>
Bug: Issue 4661