Bump scribe version to 6.9.0
The plugin classes OAuth20ServiceImpl and OAuth2AccessTokenJsonExtractor
are now replaced with the implementation from scribejava library.

New dependency on jackson-databind and jackson-annotations is required.

Bearer token authorization for request signing is supported now and is
used for providers that known to support it.

Basic client authentication is supported now and is used for providers
that known to support it.

Change-Id: I68717ebb988abe2769ccef9b6f1717bbf7ba9e28