Make preferred_username optional for Keycloak
This commit adds a new config option 'use-preferred-username' which
can be used with the Keycloak provider to make it optional to use
the preferred_username delivered by Keycloak.

This is useful in situations where the preferred_username is set
as an email address, which don't work as a gerrit username.

Leaving the username unset also allows the user to choose their
own username.

Change-Id: Iff114537527c672190c6a1a41726d2c32bb97810